src/requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.1,<3.0
djangorestframework>=3.14,<4.0